Remove dependency on ariakit
