+++ packages/bricks/src/secret-internals.ts
@@ -5,9 +5,11 @@
 
 export { Dot } from "./~utils.Dot.js";
 export { GhostAligner } from "./~utils.GhostAligner.js";
-export { DisclosureArrow } from "./~utils.icons.js";
+export { createIconFromPath, DisclosureArrow } from "./~utils.icons.js";
 export {
 	IconButtonContext,
 	IconButtonPresentation,
 } from "./IconButton.internal.js";
 export { TooltipContext } from "./Tooltip.internal.js";
+
+export type { BaseProps, FocusableProps } from "./~utils.props.js";
