+++ packages/bricks/src/~utils.icons.tsx
@@ -3,8 +3,9 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
+import { Role } from "@ariakit/react/role";
 import { Icon } from "@stratakit/foundations";
-import { createSvgFromPath, forwardRef } from "@stratakit/internal-utils";
+import { forwardRef } from "@stratakit/internal-utils";
 import cx from "classnames";
 
 import type { BaseProps } from "@stratakit/internal-utils";
@@ -12,11 +13,25 @@ import type { BaseProps } from "@stratakit/internal-utils";
 // ----------------------------------------------------------------------------
 
 /** Creates an inline icon component for the specified `<path>` definition (`d` attribute). */
-function createIconFromPath(d: string) {
-	const Svg = createSvgFromPath(d);
+export function createIconFromPath(d: string) {
 	return forwardRef<"svg", Omit<BaseProps<"svg">, "children">>(
 		(props, forwardedRef) => {
-			return <Icon render={<Svg {...props} ref={forwardedRef} />} />;
+			return (
+				<Icon
+					render={
+						<Role.svg
+							width="16"
+							height="16"
+							fill="none"
+							viewBox="0 0 16 16"
+							{...props}
+							ref={forwardedRef}
+						>
+							<path fill="currentColor" d={d} />
+						</Role.svg>
+					}
+				/>
+			);
 		},
 	);
 }
