--- packages/internal-utils/src/icons.tsx
@@ -1,35 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-import { Role } from "@ariakit/react/role";
-import { forwardRef } from "./utils.js";
-
-import type { BaseProps } from "./utils.js";
-
-// ----------------------------------------------------------------------------
-
-/** Creates an inline svg for the specified `<path>` definition (`d` attribute). */
-function createSvgFromPath(d: string) {
-	return forwardRef<"svg", Omit<BaseProps<"svg">, "children">>(
-		(props, forwardedRef) => {
-			return (
-				<Role.svg
-					width="16"
-					height="16"
-					fill="none"
-					viewBox="0 0 16 16"
-					{...props}
-					ref={forwardedRef}
-				>
-					<path fill="currentColor" d={d} />
-				</Role.svg>
-			);
-		},
-	);
-}
-
-// ----------------------------------------------------------------------------
-
-export { createSvgFromPath };
