+++ packages/internal-utils/src/utils.ts
@@ -5,9 +5,6 @@
 
 import * as React from "react";
 
-import type { FocusableProps as AkFocusableProps } from "@ariakit/react/focusable";
-import type { RoleProps } from "@ariakit/react/role";
-
 // ----------------------------------------------------------------------------
 
 const isBrowser = typeof document !== "undefined";
@@ -78,28 +75,6 @@ const forwardRef = React.forwardRef as ForwardRefHelper;
 
 // ----------------------------------------------------------------------------
 
-/** Element type props merged with custom props. */
-type MergeProps<
-	ElementType extends React.ElementType,
-	CustomProps extends Record<string, unknown>,
-> = CustomProps &
-	Omit<React.ComponentPropsWithoutRef<ElementType>, keyof CustomProps>;
-
-/** Base component props with custom props. */
-type BaseProps<ElementType extends React.ElementType = "div"> = MergeProps<
-	ElementType,
-	Pick<RoleProps, "render">
->;
-
-// ----------------------------------------------------------------------------
-
-/** Focusable component props with custom props. */
-type FocusableProps<ElementType extends React.ElementType = "div"> =
-	BaseProps<ElementType> &
-		Pick<AkFocusableProps, "disabled" | "accessibleWhenDisabled" | "autoFocus">;
-
-// ----------------------------------------------------------------------------
-
 /** See https://github.com/Microsoft/TypeScript/issues/29729 */
 type AnyString = string & {};
 
@@ -110,7 +85,7 @@ const identity = <T>(value: T) => value;
 
 // ----------------------------------------------------------------------------
 
-export type { AnyString, BaseProps, FocusableProps };
+export type { AnyString };
 
 export {
 	forwardRef,
